requirements.txt
amqp==2.3.2
anyjson==0.3.3
asn1crypto==0.24.0
billiard==3.5.0.4
boto==2.48.0
boto3==1.4.7
botocore==1.7.48
celery==4.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.11
django-allauth==0.34.0
django-anymail==1.2
django-cors-headers==2.1.0
django-debug-toolbar==1.9.1
django-enumfields==0.9.0
django-filter==1.1.0
django-rest-auth==0.9.2
django-rest-knox==3.0.3
django-storages==1.6.5
django-user-sessions==1.5.3
djangorestframework==3.4.6
docutils==0.14
factory-boy==2.9.2
Faker==0.9.0
Flask==1.0.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
kombu==4.2.1
Markdown==2.6.9
MarkupSafe==1.0
msgpack-python==0.4.8
oauthlib==2.1.0
phonenumbers==8.8.5
psycopg2==2.7.3.2
pycparser==2.18
PyJWT==1.6.4
pyOpenSSL==18.0.0
python-dateutil==2.7.3
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
raven==6.3.0
rehive==1.1.7
requests==2.18.4
requests-oauthlib==1.0.0
s3transfer==0.1.13
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
sqlparse==0.2.4
text-unidecode==1.2
urllib3==1.22
vine==1.1.4
Werkzeug==0.14.1
whitenoise==4.0
django-anymail[sendgrid]==3.0.0

